#4 Implement provider in slack
