+++ tracklater/tests/slack/test_slack.py
@@ -1,7 +1,6 @@
 from timemodules.slack import Parser
 
 import pytest
-import json
 import os
 import pytz
 
@@ -10,62 +9,6 @@ from datetime import datetime, timedelta
 DIRECTORY = os.path.dirname(os.path.realpath(__file__))
 
 HEL = pytz.timezone('Europe/Helsinki')
-
-
-@pytest.fixture(autouse=True)
-def mock_slack(monkeypatch):
-    def _api_call(*args, **kwargs):
-        if args[0] == "users.list":
-            return {
-                "members": [
-                    {
-                        "id": "1",
-                        "profile": {
-                            "first_name": "Firstname",
-                            "last_name": "Lastename"
-                        }
-                    },
-                    {
-                        "id": "2",
-                        "profile": {
-                            "first_name": "Secondname",
-                            "last_name": "Lastename"
-                        }
-                    }
-                ]
-            }
-        if args[0] == "conversations.list":
-            return {"channels": [{"id": "1"}]}
-        if args[0] == "conversations.history":
-            return {
-                "messages": [
-                    {
-                        "user": "1",
-                        "text": "First Message",
-                        "ts": "1234567"
-                    },
-                    {
-                        "user": "1",
-                        "text": "Second Message",
-                        "ts": "1234567"
-                    },
-                    {
-                        "user": "2",
-                        "text": "Third Message",
-                        "ts": "1234567"
-                    }
-                ]
-            }
-
-    class MockClient(object):
-        def __init__(self, *args):
-            pass
-
-        def api_call(self, *args, **kwargs):
-            return _api_call(*args, **kwargs)
-
-    monkeypatch.setattr('timemodules.slack.WebClient',
-                        MockClient)
 
 
 @pytest.fixture()
